update [렉스] 백준 4796 캠핑
